+++ src/bin/mmixasm.rs
@@ -1,5 +1,5 @@
 /// MMIX Assembler - Compile .mms assembly files to .mmo object code
-use checksmix::{MMixAssembler, MmoGenerator};
+use checksmix::MMixAssembler;
 use clap::Parser;
 use std::fs;
 use std::path::{Path, PathBuf};
@@ -112,10 +112,7 @@ fn main() {
     }
 
     // Generate object code
-    let object_code = MmoGenerator::new(assembler.instructions.clone(), assembler.labels.clone())
-        .with_debug_strings(assembler.debug_strings().to_vec())
-        .with_greg_inits(assembler.greg_inits.clone())
-        .generate();
+    let object_code = assembler.generate_object_code();
 
     println!("Generated {} bytes of object code", object_code.len());